pyproject.toml
[tool.poetry]
name = "big data project"
version = "0.1.0"
description = ""
authors = ["mullam55 <mullam55@students.rowan.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cassandra-driver = "^3.29.2"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}
pyyaml = "^6.0.2"
requests = "^2.32.3"
chardet = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
